% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cal.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculation for sample}
\usage{
calculate_CI(obj, conf = 0.95)
}
\arguments{
\item{conf}{number, a number of confidence interval, ranged from 0 to 1, the default is 0.95}

\item{x}{vector, a vector of numbers of length N.}
}
\value{
returns a named vector of length 2, where the first value is the lower_bound, the second value is the upper_bound.
}
\description{
Calculate confidence intervals
}
\details{
This function computes the confidence intervals for a numeric vector
}
\examples{
calculate_CI(rnorm(10,2,1))
calculate_CI(rnorm(10,2,1), 0.90)
calculate_CI(c(rnorm(10,2,1),Na), 0.95 )
}
