% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_sine.R
\name{fn_cos}
\alias{fn_cos}
\title{Calculation for cosine}
\usage{
fn_cos(x, k)
}
\arguments{
\item{x}{number, a single number to be calculated}

\item{k}{integer, an integer larger than 1}
}
\value{
the result of cosine of a number is returned.
}
\description{
Calculate cos(x)
}
\details{
This function computes the approximation to the cosine function
}
\examples{
fn_cos(5, 10)
}
