% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sine.R
\name{fn_sin}
\alias{fn_sin}
\title{Calculation for sine}
\usage{
fn_sin(x, k)
}
\arguments{
\item{x}{number, a single number to be calculated}

\item{k}{integer, an integer larger than 1}
}
\value{
the result of sine of a number is returned.
}
\description{
Calculate sin(x)
}
\details{
This function computes the approximation to the sine function
}
\examples{
fn_sin(5, 10)
}
