% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_S3.R
\name{make_ci_class}
\alias{make_ci_class}
\title{constructor}
\usage{
make_ci_class(x, conf = 0.95)
}
\arguments{
\item{x}{vector, a vector of numbers of length N.}

\item{conf}{number, a number of confidence interval, ranged from 0 to 1, the default is 0.95}
}
\value{
returns an object of class "ci_class"
}
\description{
make a constructor for ci_class object
}
\details{
This function creates a constructor function for ci_class
}
