% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Sample Mean}
\usage{
sample_mean(x)
}
\arguments{
\item{x}{vector, a vector of numbers of length N.}
}
\value{
returns the mean of the vector.
}
\description{
Internal function to calculate the sample mean.
}
\examples{
sample_mean(c(1:10))

}
