% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_variance.R
\name{sample_sd}
\alias{sample_sd}
\title{Sample Standard deviation}
\usage{
sample_sd(x)
}
\arguments{
\item{x}{vector, a vector of numbers of length N.}
}
\value{
returns the standard deviation of the vector.
}
\description{
Internal function to calculate the sample standard deviation.
}
\examples{
sample_sd(c(1:10))

}
